% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{core_file_constructor}
\alias{core_file_constructor}
\title{function to construct filenames from aws s3 core bucket}
\usage{
core_file_constructor(time, scope.orgtype, scope.formtype)
}
\arguments{
\item{time}{character vector. Dates of core/bmf files to query. Valid
inputs range from 1989-2022. Default value is "current" for 2022.}

\item{scope.orgtype}{character scalar. Organization type to query from
core/bmf s3 bucket. Valid inputs are 'CHARITIES' for charities (501C3-PC),
'PRIVFOUND' for private foundations (501C3-PF) and 'NONPROFIT' for all
nonprofits (501CE)}

\item{scope.formtype}{character scalar. Form type to query from core/bmf s3
bucket. Valid inputs are 'PC'(nonprofits that file the full version),
'EZ'(nonprofits that file 990EZs only), '
PZ'(nonprofits that file both PC and EZ), or 'PF'(private foundations).}
}
\value{
character vector. Vector of .csv filenames to query in s3 bucket.
}
\description{
This function constructs names for .csv files found in the
nccsdata bucket for nccs core files based on user inputs
}
