% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_helper.R
\name{dat_filter}
\alias{dat_filter}
\title{Function to filter data table}
\usage{
dat_filter(dat, args = args, ex_args = ex_args, id_col, census.level)
}
\arguments{
\item{dat}{data table. Data table of either census or block data.}

\item{args}{string expression. User arguments specified in parse_geo()}

\item{id_col}{string. Name of column with Tract or Block IDs}

\item{exp_args}{string expression. Processed user arguments specified in
parse_geo()}

\item{geo.level}{name of data.table (Block/Census) for error messages.}
}
\value{
A list of fips codes based on filter criteria.
}
\description{
dat_filter() takes in arguments specified by the user (name of
dataset, column names and subsets), checks if those filters apply to the
specified dataset and returns FIPs codes that meet the subsetting conditions.

A user specifies whether they are pulling data from the Block or Tract
datasets and specifies the subset of the data they want e.g. (geo.state.abbr
= c("NY")). The function returns either the Tract or Block IDs that fit
the subset specified
}
