% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dic_from_df}
\alias{dic_from_df}
\title{Function that creates a dictionary from 2 data.frame columns}
\usage{
dic_from_df(df, keycol, valcol)
}
\arguments{
\item{df}{data.frame. Data.frame to construct dictionary from}

\item{keycol}{character scalar. Name of column containing dictionary keys.}

\item{valcol}{character scalar. Name of column containing dictionary values.}
}
\value{
a dictionary with unique key-value pairs.
}
\description{
This function takes one data.frame column as a key column
and another as a value.column. It returns a dictionary mapping unique
keys to values.
}
