% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{filter_data}
\alias{filter_data}
\title{Function to filter a data.table with user-provided arguments}
\usage{
filter_data(dt, filters)
}
\arguments{
\item{dt}{data.table. Data.table to filter}

\item{filters}{list. List of column filters to apply}
}
\value{
filtered data.table
}
\description{
This function takes a list of filters and filters either the
core or bmf data.table objects.
}
