% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_helper.R
\name{fips_map}
\alias{fips_map}
\title{function that maps fips codes to user arguments}
\usage{
fips_map(geo.state, geo.city, geo.county)
}
\arguments{
\item{geo.county}{character vector. County names for filtering e.g.
"cullman", "dale". Case insensitive.}
}
\value{
character vector. county fips codes for filtering core datasets.
}
\description{
This function takes in the geographic arguments from get_data()
and maps them to fips codes found in the core datasets. These codes can
then be used to filter the core datasets.
}
