% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstupper}
\alias{firstupper}
\title{Function to process first letter of lowercase strings}
\usage{
firstupper(string)
}
\arguments{
\item{string}{character scalar. String to process and capitalize the
first letter.}
}
\value{
string with first letter capitalized
}
\description{
This function capitalizes the first character of each string.
}
