% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{generate_ntee_regex}
\alias{generate_ntee_regex}
\title{This function returns a regex query to parse NTEE2 Codes}
\usage{
generate_ntee_regex(ntee.group, ntee.code, ntee.orgtype)
}
\arguments{
\item{ntee.group}{character vector. Vector of desired Industry Group codes
to filter. Use "all" to include all possible codes.}

\item{ntee.code}{character vector. Sequence of desired Industry, Division and
Subdivision codes (old code structure) to use in filtering. Use "all" to
include all possible codes. Can also provide only partial codes. For example
"A" or "Axx" will query NTEE2 codes based on Industry group "A" and all
division and subdivisions.}

\item{ntee.orgtype}{character vector. Vector of Organization Types.
Use "all" to include all possible codes.}
}
\value{
regex query that can be used in parse_ntee() function to filter
population of NTEE2 codes.
}
\description{
This function creates the regex query that will be used
to filter NTEE2 codes based on user inputs.
}
\examples{
generate_ntee_regex("ART", "A23", "RG")
generate_ntee_regex("all", "Axx", "RG")
generate_ntee_regex("EDU", "B", "all")
}
