% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_helper.R
\name{geo_data_get}
\alias{geo_data_get}
\title{Preprocess Census Block and Tract Crosswalks}
\usage{
geo_data_get(block_s3_url, tract_s3_url)
}
\arguments{
\item{block_s3_url}{string. Path to S3 bucket with block crosswalk.}

\item{tract_s3_url}{string. Path to S3 bucket with tract crosswalk.}
}
\value{
A string message indicating that the processed data.tables
are available in memory
}
\description{
This function downloads the crosswalk datasets from an S3
bucket, reads them as data.tables, reformats the column names, and creates
a new column with state abbreviations for the Tract dataset.
}
\note{
Can also be used to recreate the data.tables if files get corrupted.
}
