% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_main.R
\name{geo_preview}
\alias{geo_preview}
\title{Preview CBSA/CSA FIPS Codes}
\usage{
geo_preview(geo, within = NULL, type = NULL)
}
\arguments{
\item{geo}{Character vector. Vector of columns to preview/return from CBSA
data}

\item{within}{character vector. Vector of state abbreviations to filter data
with.}

\item{type}{character scalar. Scalar variable indicating type of data to return.
Acceptable values are "metro" and "micro".}
}
\value{
printout of data or data.frame of filtered cbsa data.frame
}
\description{
geo_preview() previews cbsa metadata for use in filtering of
core data sets. It also saves the outputs with an invisible return.
}
\examples{

geo_preview( geo="cbsa", within="LA" )
geo_preview( geo=c("county","cbsa"), within="GA", type="metro" )
geo_preview( geo=c("cbsa","cbsafips"), within="FL", type="metro" )
xx <- geo_preview( geo=c("cbsa","cbsafips"), within="FL", type="metro" )

}
