% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_bmf}
\alias{get_bmf}
\title{Function to download master bmf file and filter it based on ntee
and FIPS codes}
\usage{
get_bmf(url, dest_path = "bmf.rds", filters)
}
\arguments{
\item{url}{character scalar. Link to object in s3 bucket.}

\item{dest_path}{character scalar. Path to download bmf file to.}

\item{filters}{list. List of column filters to apply}
}
\value{
data.table. Data.table with filtered master bmf file.
}
\description{
This function downloads an .rds file from a public s3 bucket,
reads it into memory, and deletes the file. It then converts the data.frame
into a data.table and filters it based on user-specified FIPS codes and
ntee codes.
}
