% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to query bmf data and integrate it with census, cbsa and ntee
metadata}
\usage{
get_data(ntee.level1,
                ntee.level2,
                geo.state,
                geo.metro,
                geo.level)
}
\arguments{
\item{geo.state}{string or vector. Filter query by state abbreviations e.g.
"NY", "CA". Default == NULL includes all states.}

\item{ntee.level1}{string or vector. Nonprofit Industry Group. Default ==
"all" includes all Industry Groups.}

\item{ntee.level2}{string or vector. Level 2-4 of NTEE code (Industry,
Division and Subdivision). Default == "all" includes all codes.}

\item{geo.metro}{string or vector. Filter query by cbsa code. Default = NULL
includes all metro cbsa codes.}

\item{geo.level}{string. Census dataset to merge with. Default == "tract"
which merges filtered bmf data with census tract data.}
}
\value{
data.table with queried data
}
\description{
This function uses user inputs to query, filter and merge nccs
data and additional census, cbsa and ntee metadata
}
