% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{get_ntee_level_3_4}
\alias{get_ntee_level_3_4}
\title{Function to generate level 3 and 4 of NTEE2 code}
\usage{
get_ntee_level_3_4(digits23, digits45)
}
\arguments{
\item{digits23}{character. Digits in the 2nd and 3rd place of old code}

\item{digits45}{character. Digits in the 4th and 5th place of old code}
}
\value{
The level 3 and 4 codes concatenated together in a string
}
\description{
This function takes in digits23 and digits45 from the old
NTEE codes to create levels 3 and 4 of NTEE2 code
}
\note{
See https://github.com/Nonprofit-Open-Data-Collective/mission-taxonomies/blob/main/NTEE-disaggregated/README.md
for more details
}
\examples{
\dontrun{
get_ntee_level_3_4("23", "45")
get_ntee_level_3_4("03", "22")
}

}
