% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jsonify_f}
\alias{jsonify_f}
\title{Script with utility functions
Function that prettifies json}
\usage{
jsonify_f(f)
}
\arguments{
\item{f}{json string. String containing json text}
}
\description{
Script with utility functions
Function that prettifies json
}
