% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{FUNCTION_TITLE}
\usage{
get_data(
  ntee.level1 = "all",
  ntee.level2 = "all",
  geo.state = NULL,
  geo.metro = NULL,
  geo.level = "tract"
)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{subset}{PARAM_DESCRIPTION}

\item{weights}{PARAM_DESCRIPTION}

\item{na.action}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'qr'}

\item{model}{PARAM_DESCRIPTION, Default: TRUE}

\item{x}{PARAM_DESCRIPTION, Default: FALSE}

\item{y}{PARAM_DESCRIPTION, Default: FALSE}

\item{qr}{PARAM_DESCRIPTION, Default: TRUE}

\item{singular.ok}{PARAM_DESCRIPTION, Default: TRUE}

\item{contrasts}{PARAM_DESCRIPTION, Default: NULL}

\item{offset}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{model.frame}}
}
