% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dt}
\alias{load_dt}
\title{Function to read legacy csv files from core/bmf
public s3 bucket url. Saves the file as a data.table.}
\usage{
load_dt(url)
}
\arguments{
\item{url}{character scalar. Link to public s3 object.}
}
\value{
data.table
}
\description{
This function is used in lapply to map a list of urls to a
list of data.tables.
}
