% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_main.R
\name{map_countyfips}
\alias{map_countyfips}
\title{function that maps county FIPS codes to user arguments}
\usage{
map_countyfips(geo.county, geo.region, geo.cbsafips, geo.csafips)
}
\arguments{
\item{geo.county}{character vector. County names for filtering e.g.
"cullman", "dale". Case insensitive.}

\item{geo.region}{character vector. Regions for filtering e.g. "South",
"Midwest" based on census region classifications.}

\item{geo.cbsafips}{numeric vector. Census CBSA FIPS codes.}

\item{geo.csafips}{numeric vector. Census CSA FIPS codes.}
}
\value{
character vector. county fips codes for filtering core datasets.
}
\description{
This function takes in the geographic arguments
and maps them to county FIPS codes found in legacy datasets. These codes can
then be used to filter these legacy datasets.
}
