% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{ntee_metadata}
\alias{ntee_metadata}
\title{Create metadata from NTEE data set for ntee_preview()}
\usage{
ntee_metadata(ntee.user)
}
\arguments{
\item{ntee.user}{character vector. Vector of unique NTEE2 codes.}
}
\value{
named list. Named list contains Level 1, Level 2-4, Level 5 and
processed user inputs
}
\description{
Metadata generating function for ntee_preview(). Creates
named lists for each ntee code level. Processes user inputs to ntee_preview()
for easy look up using named lists
}
