% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{ntee_preproc}
\alias{ntee_preproc}
\title{Script containing helper functions for ntee_main.R}
\usage{
ntee_preproc(path_to_csv)
}
\arguments{
\item{path_to_csv}{character. Path to the .csv file with disaggregated NTEE
data}
}
\value{
A list containing the population of valid NTEE2 codes
}
\description{
This function reads in a csv containing disaggregated NTEE
codes built from the old formatting. It uses the rules specified in
https://github.com/Nonprofit-Open-Data-Collective/mission-taxonomies/blob/main/NTEE-disaggregated/README.md
to derive the codes from level 1-5 used in NTEE2 from old NTEE codes
}
\note{
The disaggregated csv file can be edited and the function rerun
to generate a new list of valid NTEE2 codes if needed.
}
