% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_main.R
\name{ntee_preview}
\alias{ntee_preview}
\title{Function to document NTEE Codes}
\usage{
ntee_preview(ntee, ntee.group, ntee.code, ntee.orgtype)
}
\arguments{
\item{ntee}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Vector of desired Industry Group codes
to filter. Use "all" to include all possible codes.}

\item{ntee.code}{character vector. Sequence of desired Industry, Division and
Subdivision codes (old code structure) to use in filtering. Use "all" to
include all possible codes. Can also provide only partial codes. For example
"A" or "Axx" will query NTEE2 codes based on Industry group "A" and all
division and subdivisions.}

\item{ntee.orgtype}{character vector. Vector of Organization Types.
Use "all" to include all possible codes.}
}
\value{
text strings that describe NTEE2 codes at each level
}
\description{
This function takes user inputs across levels 1- 5 of the
NTEE2 code structure and returns text metadata describing each NTEE Code
level
}
\details{
Script with front end functions for users to access NTEE2 codes and
filter NTEE dataset with codes
}
\examples{
ntee_preview(ntee = c("ART", "A2X"))

}
