% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{nteecc_map}
\alias{nteecc_map}
\title{This function maps ntee2 codes to ntee.cc codes}
\usage{
nteecc_map(ntee.user, ntee.group, ntee.code, ntee.orgtype)
}
\arguments{
\item{ntee.user}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Specific Industry Group codes submitted
by user}

\item{ntee.code}{character vector. Specific level 2-4 codes (Industry,
Division, Subdivision) submitted by user.}

\item{ntee.orgtype}{character vector. Specific level 5 codes (Organization
Type) submittted by user.}
}
\value{
character vector. ntee.cc codes for merging in the core dataset.
}
\description{
This function takes as an input ntee2 codes inferred from
user inputs in get_data() and maps them to ntee.cc codes for merging
found in ntee_df
}
