% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obj_validate}
\alias{obj_validate}
\title{Function to validate constructed file names and return s3 keys}
\usage{
obj_validate(dsname, filenames, bucket.str, base.url, return.key)
}
\arguments{
\item{dsname}{character scalar. Name of data series to query from S3.
Valid inputs are "core" and "bmf", not both.}

\item{filenames}{character vector. Vector of file names returned by
core_file_constructor}

\item{bucket.str}{character scalar. Url of legacy core/bmf bucket for
string formatting}

\item{base.url}{character scalar. Base url of nccsdata s3 bucket.}

\item{return.key}{boolean. Default == FALSE to return url links to s3
buckets. Default == TRUE returns bucket keys for s3_select queries.}
}
\value{
vector of valid s3 bucket keys for core data sets
}
\description{
This function takes constructed file names from
core_file_constructor() and tests if the urls to those files exist. Then it
extracts s3 bucket keys from the validated urls for downstream s3 queries or
returns the validated urls themselves for local download.
}
