% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{objs_exist}
\alias{objs_exist}
\title{Check if ALL objects named one vector exist in memory}
\usage{
objs_exist(c("block_dt", "tract_dt"))
}
\arguments{
\item{...}{vector of objects.}
}
\value{
A single boolean value indicating whether all objects are present 
in memory
}
\description{
This function is used to check if both block and tract data
tables exist in memory
}
