% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_helper.R
\name{parse_fips}
\alias{parse_fips}
\title{Function that returns FIPS codes that match dynamic User arguments}
\usage{
parse_fips(census.level, ...)
}
\arguments{
\item{census.level}{string. data.table to parse; "BLOCK" | "TRACT"}

\item{...}{columns in either tract or block dataset for filtering.}
}
\value{
a list of FIPS codes for either Tract IDs or Block IDs.
}
\description{
Filters either the Block or Tract data.tables to return
a list of FIPS codes that match conditions specified by the User
}
\examples{
\dontrun{
parse_fips(census.level = "TRACT",
           state.census.abbr = c("NY", "MD"))
}

}
