% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{parse_ntee_regex}
\alias{parse_ntee_regex}
\title{Function to loop through and execute a vector of regex queries
on a vector of ntee codes}
\usage{
parse_ntee_regex(regexp_vec, ntee_codes)
}
\arguments{
\item{regexp_vec}{character vector. Vector containing all regex queries}

\item{ntee_codes}{character vector. Vector containing population of NTEE2
codes}
}
\value{
vector of matched NTEE2 codes
}
\description{
This function executes a vector of regex queries on
a vector containing the population of NTEE2 codes
}
