% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{paws_s3_select}
\alias{paws_s3_select}
\title{function to perform s3_select with paws}
\usage{
paws_s3_select(bucket, key, query, file.header, csv.header)
}
\arguments{
\item{bucket}{character scalar. Name of s3 bucket.}

\item{key}{character scalar. s3 object key.}

\item{query}{character scalar. Desired SQL query.}

\item{file.header}{character scalar. Option to use header columns in SQL
query. Acceptable inputs are USE | NONE | IGNORE}

\item{csv.header}{boolean. Option to return csv with first row as header.}
}
\value{
dataframe with query results
}
\description{
This function uses paws to perform aws s3 select queries on a
user provided object key and bucket.
}
