% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_meta.R
\name{preview_meta}
\alias{preview_meta}
\title{This function filters the cbsa dataframe.}
\usage{
get_arg_values(dataset, visual)
}
\arguments{
\item{dataset}{string. Name of dataset to load, "cbsa", "block", "tract".}

\item{visual}{boolean. Option to return reactable visualization or filtered
dataframe}

\item{within}{character vector. Vector of column variables to filter by
without explicit argument definition, filters all rows with columns
containing any of the within arguments. For example within = c("NY",
"Alabama") will return all rows containing either NY or Alabama}

\item{...}{expression. User inputs of selected columns and values to filter
by. E.g. (state.census.abbr = c("NY", "AL")). Leaving blank returns all
columns}
}
\value{
filtered dataframe or list with filtered dataframe and first 20
rows of table visualized with reactable
}
\description{
This function takes in a cbsa dataset and a dynamic list
of user selected columns. It returns a filtered dataframes based on user
conditions
}
\examples{
get_arg_values("cbsa", TRUE, state.census.name = c("Wyoming", "Montana"))
get_arg_values("tract", TRUE,
                metro.census.cbsa.geoid = c("10100", "10200"),
                state.census.abbr = c("NY", "CA"))
get_arg_values("tract", TRUE, within = ("NY", "Alabama"))

}
