% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_sample.R
\name{preview_sample}
\alias{preview_sample}
\title{Function to summarize NCCS data.}
\usage{
preview_sample(
  data,
  group_by,
  var,
  stats,
  ntee = NULL,
  ntee.group = NULL,
  ntee.code = NULL,
  ntee.orgtype = NULL,
  geo.state = NULL,
  geo.city = NULL,
  geo.region = NULL,
  geo.county = NULL
)
}
\arguments{
\item{data}{data.frame or data.table. In-memory dataset to summarize}

\item{group_by}{character vector. Vector of columns for dplyr::group_by()}

\item{var}{character scalar. Column to calculate summary statistics with}

\item{ntee}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Specific Industry Group codes submitted
by user}

\item{ntee.code}{character vector. Specific level 2-4 codes (Industry,
Division, Subdivision) submitted by user.}

\item{ntee.orgtype}{character vector. Specific level 5 codes (Organization
Type) submitted by user.}

\item{geo.state}{character vector. Filter query by state abbreviations e.g.
"NY", "CA". Default == NULL includes all states.}

\item{geo.city}{character vector. City names for filtering e.g. "Chicago",
"montgomery". Case insensitive}

\item{geo.region}{character vector. Regions for filtering e.g. "South",
"Midwest" based on census region classifications.}

\item{geo.county}{character vector. County names for filtering e.g.
"cullman", "dale". Case insensitive.}

\item{stats.}{character vector. Vector of summary statistics to compute with
dplyr::summarise(). Available options are count, min, max, median and mean}
}
\value{
dataframe with summary statistics computed for each group
}
\description{
This function computes summary statistics for data pulled with
get_data()
}
\examples{
\dontrun{
core <- get_data(dsname = "core",
                 time = "2005")
preview_sample(data = core,
               group_by = c("NTEECC", "STATE"),
               var = c("TOTREV"),
               stats = c("count", "mean", "max"))
}

}
