% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_select.R
\name{query_construct}
\alias{query_construct}
\title{function to construct SQL queries for s3 select}
\usage{
query_construct(filters)
}
\arguments{
\item{filters}{list. List of column filters to apply}
}
\value{
a completed SQL query
}
\description{
This function takes in a list of filter arguments from
get_data() and constructs a SQL query for use in s3 select
}
