% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{query_ntee}
\alias{query_ntee}
\title{function to match user queries to ntee2 codes}
\usage{
query_ntee(ntee.user, ntee.group, ntee.code, ntee.orgtype)
}
\arguments{
\item{ntee.user}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Specific Industry Group codes submitted
by user}

\item{ntee.code}{character vector. Specific level 2-4 codes (Industry,
Division, Subdivision) submitted by user.}

\item{ntee.orgtype}{character vector. Specific level 5 codes (Organization
Type) submittted by user.}
}
\value{
a vector of matched unique ntee2 codes
}
\description{
This function runs seperate ntee queries on user inputs and
combines the results
}
