% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_select.R
\name{s3_query}
\alias{s3_query}
\title{Function to perform S3 Select query on s3 bucket}
\usage{
s3_query(bucket, keys, filters)
}
\arguments{
\item{bucket}{character scalar. Name of s3 bucket to query}

\item{keys}{character vector. s3 object keys}

\item{filters}{list. List of column filters to apply}
}
\value{
list of queried dataframes. One for each key supplied.
}
\description{
This function queries multiple core/bmf .csv S3 objects and
returns a list of object subsets that satisfy user-specified filter
conditions.
}
