% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{s3_validate}
\alias{s3_validate}
\title{Function to validate constructed file names and return s3 keys}
\usage{
s3_validate(dsname, filenames)
}
\arguments{
\item{dsname}{character scalar. Name of data series to query from S3.
Valid inputs are "core" and "bmf", not both.}

\item{filenames}{character vector. Vector of file names returned by
core_file_constructor}
}
\value{
vector of valid s3 bucket keys for core data sets
}
\description{
This function takes constructed file names from
core_file_constructor() and tests if the urls to those files exist. Then it
extracts s3 bucket keys from the validated urls for downstream s3 queries.
}
