% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{sort_ntee}
\alias{sort_ntee}
\title{Function to sort user ntee inputs into group, code and organization
tyoe}
\usage{
sort_ntee(ntee.user, ntee.group, ntee.code, ntee.orgtype)
}
\arguments{
\item{ntee.user}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Specific Industry Group codes submitted
by user}

\item{ntee.code}{character vector. Specific level 2-4 codes (Industry,
Division, Subdivision) submitted by user.}

\item{ntee.orgtype}{character vector. Specific level 5 codes (Organization
Type) submittted by user.}
}
\value{
a list with all ntee groups, codes and organization types being
queried by the user
}
\description{
This function takes as input the ntee= argument from get_data()
and sorts user inputs into group, code and organization type. It then
combines these user inputs with any pre-existing group, code and organization
types.
}
\examples{
\dontrun{
    sort_ntee(ntee.user = c("HEA", "B", "A2x", "C45", "RG", "AA"),
              ntee.group = c("ART", "EDU"))
    sort_ntee(ntee.user = c("HEA", "B", "A2x", "C45", "RG"),
              ntee.code = c("A2x", "Bxx"))
    sort_ntee(ntee.user = c("HEA", "B", "A2x", "C45"),
              ntee.orgtype = c("AA", "RG"))
}
}
