% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{validate_get_data}
\alias{validate_get_data}
\title{Function to validate inputs to get_data()}
\usage{
validate_get_data(dsname, time, scope.orgtype, scope.formtype)
}
\arguments{
\item{dsname}{character scalar. Name of data series to query from S3.
Valid inputs are "core" and "bmf", not both.}

\item{time}{character vector. Dates of core/bmf files to query. Valid
inputs range from 1989-2022. Default value is "current" for 2022.}

\item{scope.orgtype}{character scalar. Organization type to query from
core/bmf s3 bucket. Valid inputs are 'CHARITIES' for charities (501C3-PC),
'PRIVFOUND' for private foundations (501C3-PF) and 'NONPROFIT' for all
nonprofits (501CE)}

\item{scope.formtype}{character scalar. Form type to query from core/bmf s3
bucket. Valid inputs are 'PC'(nonprofits that file the full version),
'EZ'(nonprofits that file 990EZs only), '
PZ'(nonprofits that file both PC and EZ), or 'PF'(private foundations).}
}
\value{
message describing the data that is being queried.
}
\description{
This function validates each argument passed to get_data() and
returns informative error messages if user inputs are incorrect.
}
