% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_helper.R
\name{validate_inp}
\alias{validate_inp}
\title{Function to inspect user inputs and flag errors}
\usage{
validate_inp(ntee.group, ntee.code, ntee.orgtype, ind_group_codes,
                    level_2_4_codes, org_type_codes)
}
\arguments{
\item{ntee.group}{character. 3-character alphabetical Industry Group code.}

\item{ntee.code}{character. 3-character alphanumeric containing
Industry, Division and Subdivision}

\item{ntee.orgtype}{character. 2-character alphabetical Organization Type
code.}

\item{ind_group_codes}{character vector of all acceptable Industry Group
codes.}

\item{level_2_4_codes}{character vector of all acceptable Industry,
Division and Subdivision codes.}

\item{org_type_codes}{character vector of all acceptable Organization
Type codes.}
}
\value{
String indicating whether input is valid or invalid. If invalid,
points user to a list of acceptable codes.
}
\description{
This function validates user inputs for Industry Group,
Industry, Division, Subdivision and Organization Type, comparing them
to the set of codes in the population. It throws informative error
messages.
}
