% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_sample_helper.R
\name{validate_preview}
\alias{validate_preview}
\title{Function to validate inputs to preview_sample()}
\usage{
validate_preview(df_cols, group_by, var, stats)
}
\arguments{
\item{df_cols}{character vector. Vector of columns in 'data' arg in
preview_sample(). Used to ensure that query columns are present in data}

\item{group_by}{character vector. Vector of columns for dplyr::group_by()}

\item{var}{character scalar. Column to calculate summary statistics with}

\item{stats}{character vector. Vector of summary statistics to compute with
dplyr::summarise(). Available options are count, min, max, median and mean}
}
\value{
Informative error messages for the user to edit function arguments
}
\description{
This function stops execution of preview_sample() if invalid
query columns are entered
}
