% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to query bmf data and integrate it with census, cbsa and ntee
metadata}
\usage{
get_data(
  dsname = NULL,
  time = "2015",
  scope.orgtype = "NONPROFIT",
  scope.formtype = "PZ",
  geo.state = NULL,
  geo.city = NULL,
  geo.county = NULL,
  ntee = NULL,
  ntee.group = NULL,
  ntee.code = NULL,
  ntee.orgtype = NULL,
  aws = FALSE
)
}
\arguments{
\item{dsname}{character scalar. Name of data series to query from S3.
Valid inputs are "core" and "bmf", not both.}

\item{time}{character vector. Dates of core/bmf files to query. Valid
inputs range from 1989-2022.}

\item{scope.orgtype}{character scalar. Organization type to query from
core/bmf s3 bucket. Valid inputs are 'CHARITIES' for charities (501C3-PC),
'PRIVFOUND' for private foundations (501C3-PF) and 'NONPROFIT' for all
nonprofits (501CE)}

\item{scope.formtype}{character scalar. Form type to query from core/bmf s3
bucket. Valid inputs are 'PC'(nonprofits that file the full version),
'EZ'(nonprofits that file 990EZs only), '
PZ'(nonprofits that file both PC and EZ), or 'PF'(private foundations).}

\item{geo.state}{character vector. Filter query by state abbreviations e.g.
"NY", "CA". Default == NULL includes all states.}

\item{geo.city}{character vector. City names for filtering e.g. "Chicago",
"montgomery". Case insensitive}

\item{geo.county}{character vector. County names for filtering e.g.
"cullman", "dale". Case insensitive.}

\item{ntee}{character vector. Vector of user inputs. The user inputs are
progressively filtered until group, code and orgtypes are sorted into
separate vectors.}

\item{ntee.group}{character vector. Specific Industry Group codes submitted
by user}

\item{ntee.code}{character vector. Specific level 2-4 codes (Industry,
Division, Subdivision) submitted by user.}

\item{ntee.orgtype}{character vector. Specific level 5 codes (Organization
Type) submitted by user.}

\item{aws}{boolean. Whether to use aws.s3::s3_select() in executing queries.
Default == FALSE, select TRUE to use s3_select. Must have aws account to use.}
}
\value{
data.table with queried data
}
\description{
This function uses user inputs to query, filter and merge nccs
data and additional census, cbsa and ntee metadata
}
