% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntee_main.R
\name{ntee_preview}
\alias{ntee_preview}
\title{Script with front end functions for users to access NTEE2 codes and
filter NTEE dataset with codes}
\usage{
ntee_preview(ntee.group, ntee.code, ntee.orgtype, cols, visualize)
}
\arguments{
\item{ntee.group}{character vector. Vector of desired Industry Group codes
to filter. Use "all" to include all possible codes.}

\item{ntee.code}{character vector. Sequence of desired Industry, Division and
Subdivision codes (old code structure) to use in filtering. Use "all" to
include all possible codes. Can also provide only partial codes. For example
"A" or "Axx" will query NTEE2 codes based on Industry group "A" and all
division and subdivisions.}

\item{ntee.orgtype}{character vector. Vector of Organization Types.
Use "all" to include all possible codes.}

\item{cols}{character vector. A vector containing list of columns to select.
Default value is "all" for all columns.}

\item{visualize}{boolean. A boolean value indicating whether to return a
reactable object or data frame.}
}
\value{
a dataframe or reactable object with rows that match user arguments.
}
\description{
This function takes user inputs across levels 1- 5 of the
NTEE2 code structure and returns selected columns from the NTEE database.
The use can choose to return a dataframe or reactable object.
}
\details{
Function to interactively explore NTEE codes
}
\examples{
ntee_preview("all", "A1x", "all", "all",TRUE)
ntee_preview("ART", "Axx", "all", "all",TRUE)
ntee_preview("all", "Axx", "all", c("type.org", "univ"), FALSE)

}
