% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_geo_helper.R
\name{parse_geo}
\alias{parse_geo}
\title{Simplified Geo Parser for Census Crosswalks}
\usage{
parse_geo(
  census.level,
  geo.region = NULL,
  geo.state = NULL,
  geo.county = NULL,
  geo.city = NULL
)
}
\arguments{
\item{census.level}{character scalar. Name of crosswalk, either 'tract' or
'block'.}

\item{geo.region}{character vector. Name of geographic region. Acceptable
values are 'South', 'West', 'Northeast' and 'Midwest'. Default == NULL.}

\item{geo.state}{character vector. 2 letter state abbreviations. Values such
as 'NY', 'CA', 'WY' are accepted. Default == NULL.}

\item{geo.county}{character vector. Name of county. Case insensitive.
Default == NULL.}

\item{geo.city}{character vector. Name of city. State should be entered
in geo.state. Case insensitive. Default == NULL.}
}
\value{
a character vector of either census tract or block FIPS.
}
\description{
This function takes in simplified user inputs, filters
either the tract or block datasets based on these inputs and returns the
relevant FIPS codes
}
\details{
All arguments are nested. Hence only the FIPs that are present in
all filters are returned.
}
\examples{
\dontrun{
parse_geo(census.level = "tract",
          geo.region = "West",
          geo.state = "WA",
          geo.county = "king")
}

}
