% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{query_construct}
\alias{query_construct}
\title{function to construct SQL queries for s3 select}
\usage{
query_construct(geo.state, ntee)
}
\arguments{
\item{geo.state}{character vector. Vector of state abbreviations}

\item{ntee}{character vector. Vector of ntee codes found in NTEECC}
}
\value{
a completed SQL query
}
\description{
This function takes in filter arguments from get_data() and
constructs a SQL query for s3 select
}
