% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_helpers.R
\name{s3_query}
\alias{s3_query}
\title{Function to perform S3 Select query on s3 bucket}
\usage{
s3_query(bucket, keys, geo.state, ntee)
}
\arguments{
\item{bucket}{character scalar. Name of s3 bucket to query}

\item{keys}{character vector. s3 object keys}

\item{geo.state}{character vector. Vector of state abbreviations}

\item{ntee}{character vector. Vector of ntee codes found in NTEECC}
}
\value{
list of queried dataframes. One for each key supplied.
}
\description{
This function queries multiple core/bmf .csv S3 objects and
returns a list of object subsets that satisfy user-specified filter
conditions.
}
